fix redundant code bug
